requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv


sentence-transformers
pydantic
python-multipart
aiofiles
httpx
pytest
